requirements.txt
# Основные зависимости для генератора комиксов

# CrewAI и LLM
crewai>=0.28.0
litellm>=1.35.0
langchain-community>=0.0.37

# Обработка изображений
Pillow>=10.0.0
requests>=2.31.0

# Обработка PDF
pypdf>=3.15.0

# Утилиты
python-dotenv>=1.0.0

# Веб-компоненты (опционально, для веб-интерфейса)
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.0

# Разработка и тестирование
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Логирование и мониторинг
structlog>=23.1.0
rich>=13.5.0